requirements.txt
anyio==3.4.0
certifi==2021.10.8
charset-normalizer==2.0.8
Deprecated==1.2.13
googleapis-common-protos==1.53.0
grpcio==1.42.0
h11==0.12.0
httpcore==0.14.3
httpx==0.21.1
idna==3.3
opentelemetry-api==1.7.1
opentelemetry-exporter-jaeger==1.7.1
opentelemetry-exporter-jaeger-proto-grpc==1.7.1
opentelemetry-exporter-jaeger-thrift==1.7.1
opentelemetry-instrumentation==0.26b1
opentelemetry-instrumentation-httpx==0.26b1
opentelemetry-sdk==1.7.1
opentelemetry-semantic-conventions==0.26b1
protobuf==3.19.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
thrift==0.15.0
wrapt==1.13.3
